requirements.txt
Flask==3.1.3
numpy==2.2.3
Pillow==11.1.0
